Move buildscript test in integration test project
