+++ integration-tests/gradle/src/test/java/io/quarkus/gradle/buildfiles/GroovyBuildFileTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.gradle;
+package io.quarkus.gradle.buildfiles;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -7,14 +7,15 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeAll;
 
 import io.quarkus.bootstrap.model.AppArtifactCoords;
+import io.quarkus.gradle.GroovyBuildFileFromConnector;
 
-class KotlinBuildFileTest extends AbstractBuildFileTest {
+class GroovyBuildFileTest extends AbstractBuildFileTest {
 
-    private static KotlinBuildFileFromConnector buildFile;
+    private static GroovyBuildFileFromConnector buildFile;
 
     @BeforeAll
     public static void beforeAll() throws URISyntaxException, IOException {
-        buildFile = initializeProject("gradle-kts-project", KotlinBuildFileFromConnector::new);
+        buildFile = initializeProject("buildfiles/gradle-project", GroovyBuildFileFromConnector::new);
     }
 
     @Override
