+++ integration-tests/gradle/src/test/java/io/quarkus/gradle/buildfiles/KotlinBuildFileTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.gradle;
+package io.quarkus.gradle.buildfiles;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -7,14 +7,15 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeAll;
 
 import io.quarkus.bootstrap.model.AppArtifactCoords;
+import io.quarkus.gradle.KotlinBuildFileFromConnector;
 
-class GroovyBuildFileTest extends AbstractBuildFileTest {
+class KotlinBuildFileTest extends AbstractBuildFileTest {
 
-    private static GroovyBuildFileFromConnector buildFile;
+    private static KotlinBuildFileFromConnector buildFile;
 
     @BeforeAll
     public static void beforeAll() throws URISyntaxException, IOException {
-        buildFile = initializeProject("gradle-project", GroovyBuildFileFromConnector::new);
+        buildFile = initializeProject("buildfiles/gradle-kts-project", KotlinBuildFileFromConnector::new);
     }
 
     @Override
