+++ integration-tests/gradle/src/test/resources/buildfiles/gradle-kts-project/build.gradle.kts
