+++ integration-tests/gradle/src/test/resources/buildfiles/gradle-kts-project/settings.gradle.kts
