+++ integration-tests/gradle/src/test/resources/buildfiles/gradle-project/gradle.properties
