+++ integration-tests/gradle/src/test/resources/buildfiles/gradle-project/settings.gradle
